Remove stateful logic from the Creator utilities (#116)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This work consolidates logic back into the `Creator.tsx` file and only uses stateless logic in the abstract utilities file as a means to keep track of class names.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check the site still works without runtime errors, no logic has been updated, only moved around.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps
- Split areas of the Creator into their own components
